+++ include/type/git/vpg_git_form_property.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-// @@Form
+// @@Form @@Json
 enum class VPGGitFormProperty
 {
-    Log // GETSET_SPTR_NULL(VPGGitLog, Log)
+    Log // GETSET_SPTR_NULL(VPGGitLog, Log) @@NoJson
     , GitManager // MANAGER_SPTR_NULL(GitManager, GitManager, _LogConfig)
 };