+++ include/type/workspace/vpg_main_form_property.hpp
@@ -1,10 +1,8 @@
 #pragma once
 
-// @@Form
+// @@Form @@Json
 enum class VPGMainFormProperty
 {
     WorkspaceForms // VECTOR_SPTR(VPGWorkspaceForm, WorkspaceForms)
-    , CurrentWorkspaceForm // GETSET_SPTR_NULL(VPGWorkspaceForm, CurrentWorkspaceForm)
-    , AddWorkspaceForm // ACTION(AddWorkspaceForm) @@NoHistory
-    , DeleteWorkspaceForm // ACTION(DeleteWorkspaceForm) @@NoHistory
+    , Initialize // ACTION(Initialize) @@NoHistory
 };