+++ include/vpg_global.hpp
@@ -25,6 +25,10 @@ class VPGGlobal
         static std::wstring GetCppDefaultIncludePathLinux();
         static std::wstring GetCppDefaultIncludePathMacOs();
 
+        // VPM - VCC Project Manager
+        static std::wstring GetVCCProjectManagerConfigFileName();
+        static std::wstring GetVCCProjectManagerConfigFileFullPath();
+
         // Generator
         static std::shared_ptr<VPGEnumClassReader> GetEnumClassReader();
         static std::shared_ptr<VPGFileGenerationManager> GetFileGenerationManager();