+++ src/form/git/vpg_git_form.cpp
@@ -3,10 +3,17 @@
 
 #include <assert.h>
 #include <memory>
+#include <string>
 
 #include "base_form.hpp"
 #include "exception_macro.hpp"
+#include "i_document.hpp"
+#include "i_document_builder.hpp"
+#include "json.hpp"
+#include "number_helper.hpp"
+#include "string_helper.hpp"
 #include "vpg_git_form_property.hpp"
+#include "vpg_git_log.hpp"
 
 // <vcc:customHeader sync="RESERVE" gen="RESERVE">
 // </vcc:customHeader>
@@ -26,6 +33,25 @@ std::shared_ptr<IObject> VPGGitForm::Clone() const
     auto obj = std::make_shared<VPGGitForm>(*this);
     obj->CloneLog(this->_Log.get());
     return obj;
+}
+
+std::shared_ptr<Json> VPGGitForm::ToJson() const
+{
+    TRY
+        NamingStyle namestyle = NamingStyle::PascalCase;
+        auto json = std::make_unique<Json>();
+        return json;
+    CATCH
+    return nullptr;
+}
+
+void VPGGitForm::DeserializeJson(std::shared_ptr<IDocument> document) const
+{
+    TRY
+        NamingStyle namestyle = NamingStyle::PascalCase;
+        auto json = std::dynamic_pointer_cast<Json>(document);
+        assert(json != nullptr);
+    CATCH
 }
 
 void VPGGitForm::InitializeComponents() const