+++ src/form/workspace/vpg_workspace_form.cpp
@@ -3,9 +3,16 @@
 
 #include <assert.h>
 #include <memory>
+#include <string>
 
 #include "base_form.hpp"
 #include "exception_macro.hpp"
+#include "i_document.hpp"
+#include "i_document_builder.hpp"
+#include "json.hpp"
+#include "number_helper.hpp"
+#include "string_helper.hpp"
+#include "vpg_git_form.hpp"
 #include "vpg_workspace_form_property.hpp"
 
 // <vcc:customHeader sync="RESERVE" gen="RESERVE">
@@ -26,6 +33,45 @@ std::shared_ptr<IObject> VPGWorkspaceForm::Clone() const
     auto obj = std::make_shared<VPGWorkspaceForm>(*this);
     obj->CloneGitForms(this->_GitForms);
     return obj;
+}
+
+std::shared_ptr<Json> VPGWorkspaceForm::ToJson() const
+{
+    TRY
+        NamingStyle namestyle = NamingStyle::PascalCase;
+        auto json = std::make_unique<Json>();
+        // Name
+        json->AddString(ConvertNamingStyle(L"Name", NamingStyle::PascalCase, namestyle), _Name);
+        // GitForms
+        auto tmpGitForms = std::make_shared<Json>();
+        json->AddArray(ConvertNamingStyle(L"GitForms", NamingStyle::PascalCase, namestyle), tmpGitForms);
+        for (auto const &element : _GitForms) {
+            tmpGitForms->AddArrayObject(element->ToJson());
+        }
+        return json;
+    CATCH
+    return nullptr;
+}
+
+void VPGWorkspaceForm::DeserializeJson(std::shared_ptr<IDocument> document) const
+{
+    TRY
+        NamingStyle namestyle = NamingStyle::PascalCase;
+        auto json = std::dynamic_pointer_cast<Json>(document);
+        assert(json != nullptr);
+        // Name
+        if (json->IsContainKey(ConvertNamingStyle(L"Name", namestyle, NamingStyle::PascalCase)))
+            _Name = json->GetString(ConvertNamingStyle(L"Name", namestyle, NamingStyle::PascalCase));
+        // GitForms
+        ClearGitForms();
+        if (json->IsContainKey(ConvertNamingStyle(L"GitForms", namestyle, NamingStyle::PascalCase))) {
+            for (auto const &element : json->GetArray(ConvertNamingStyle(L"GitForms", namestyle, NamingStyle::PascalCase))) {
+                auto tmpGitForms = std::make_shared<VPGGitForm>();
+                tmpGitForms->DeserializeJson(element->GetArrayElementObject());
+                InsertGitForms(tmpGitForms);
+            }
+        }
+    CATCH
 }
 
 void VPGWorkspaceForm::InitializeComponents() const